refactor(core): split engine internals
