+++ .changeset/shiny-dodos-act.md
@@ -1,0 +1,4 @@
+---
+---
+
+Refactor core engine internals without changing published package behavior.
